backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.11
Pillow==10.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
dj-database-url==3.1.2
django-filter==24.1
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
openai>=1.40.0
pgvector>=0.3.0
google-generativeai>=0.8.3
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
gradio-client>=1.0.0
